Implemented react bootstrap color scheme to help classify each project and refactored the renderProgressBar function to utilize react-bootstrap progress bars instead of ramonak. Added grandparent and parent elements, table and tbody, to the render data table function. Refactored the transfer info and corresponding transfer all button to be render in a function that is called only if the showTransferInfo props is true.
In the selectors component added allProjects, data and projects keys to the formattedProjects, refundInfo and myPendingTransfers return value objects respectively. Used reject instead of without to fetch closed projects in the MyClosedProjectsSelector because reject is better suited for filtering through collections. Modified the selectors such that they only accept functions, not object paths of the return values of functions.