Updated bootstrap dependencies to latest versions. Changed the width of the main card components to where they are all 1/3 the width of the screen. Imported the MyContributions component to the Content component. Added a project popover feature that displays whenever the user hovers over a particular contribution table row. The project displayed is the project associated with the contribution.
Added a refund progress bar to the cancelled and failed projects. Merged the renderProgressBar and renderDataTable functions into the renderContent function. Every displayed project that accepts individual input, like the cancel button for my open projects or the contribute form for the discover projects, is now the parent of a <li></li> element listed below the projects content.

Added the renderRefundAllButton function to the pendingRefund tab, which displays only when the fee percent has loaded and the refundTotal is greater than one. Changed the name of the myPendingContributionRefundsSelector to myPendingRefundsSelector to shorten the length of the variable name.

In the MyProjects component, I added the showTransferAllButton and removed the showTransferInfo button which only renders when the totalPendingTransferFunds is greater than 0. Added an overlay trigger to the button which displays the Fee and Total amount within a tool tip instead of in a set of div elements below the button.

Modified the CONTRIBUTION_REFUNDED action case in the reducers file such that the findIndex function looks for refunds already store that are both equal to the action refund.id (refund.id is equal to the project id the refund is associated with) and refund.supporter instead of refund.id alone.

Removed the unnecessary account parameter to the myPendingRefundsSelector. Added a refundTotal key to the myPendingRefunds object.